acmicpc/src/test/java/problem/step07/Q1157Test.java
package problem.step07;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;

class Q1157Test {

    @ParameterizedTest(name = "{index} [{arguments}]")
    @MethodSource
    void test(String enter, String expected) throws Exception {
        assertThat(Q1157.solution(enter)).isEqualTo(expected);
    }

    private static Stream<Arguments> test() {
        return Stream.of(
                Arguments.of("Mississipi", "?"),
                Arguments.of("zZa", "Z"),
                Arguments.of("z", "Z"),
                Arguments.of("baaa", "A")
        );
    }
}